update checker & fix some db issue

+++ src/main/java/dev/isnow/allahfinder/checker/protocol/packet/impl/HandshakePacket.java
@@ -3,7 +3,6 @@ package dev.isnow.allahfinder.checker.protocol.packet.impl;
 import dev.isnow.allahfinder.checker.connection.ConnectAtributes;
 import dev.isnow.allahfinder.checker.protocol.packet.Packet;
 import dev.isnow.allahfinder.util.PacketUtil;
-
 import java.io.DataOutputStream;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -13,9 +12,9 @@ public class HandshakePacket extends Packet {
 
     private final String ip;
     private final int protocol;
-    private final short port;
+    private final int port;
 
-    public HandshakePacket(DataOutputStream wrappedStream, ArrayList<ConnectAtributes> connectAtributes, String name, String ip, int protocol, short port) {
+    public HandshakePacket(DataOutputStream wrappedStream, ArrayList<ConnectAtributes> connectAtributes, String name, String ip, int protocol, int port) {
         super("Handshake", wrappedStream, connectAtributes);
         this.ip = ip;
         this.protocol = protocol;
@@ -25,6 +24,7 @@ public class HandshakePacket extends Packet {
     @Override
     public void run() throws IOException {
         PacketUtil.writeByte(outputStream, 0x00);
+
         PacketUtil.writeVarInt(outputStream, protocol);
 
         if(!connectAtributes.contains(ConnectAtributes.PREMIUM) || connectAtributes.contains(ConnectAtributes.COCKED)) {
@@ -33,7 +33,7 @@ public class HandshakePacket extends Packet {
             PacketUtil.writeString(outputStream, ip);
         }
 
-        PacketUtil.writeShort(outputStream, port);
+        PacketUtil.writeShort(outputStream, (short) port);
         PacketUtil.writeVarInt(outputStream, 2);
     }
 }
